WP4_Thor-main/service/crawlers/WebClassifier/requirements.txt
kafka-python
iocextract
elasticsearch
torch
nltk==3.7
